+++ packages/vital-layout/src/components/Footer/tokens/vitalFooter.ts
@@ -50,7 +50,7 @@ const Base = asFooterToken({
     Wrapper: vitalColor.BgSecondaryFooter,
   },
   Schema: {
-    FooterMenu: withNodeKey({ nodeKey: 'footer-menu', nodeCollection: 'site' }),
+    FooterMenu: withNodeKey('footer-menu'),
     _: withNode,
   },
 });