+++ packages/vital-templates/src/components/GenericTemplate/tokens/vitalGenericTemplate.ts
@@ -16,7 +16,6 @@ import {
   on,
   as,
   flowIf,
-  addProps,
   withDesign,
   Img,
   Fragment,
@@ -29,7 +28,6 @@ import { vitalFlowContainer } from '@bodiless/vital-flowcontainer';
 import { ContentListingClean, vitalContentListing } from '@bodiless/vital-content-listing';
 import { withNode, withNodeKey, useNode } from '@bodiless/core';
 import { vitalSpacing } from '@bodiless/vital-elements';
-import { SearchLayoutClean, vitalSearchLayout } from '@bodiless/vital-search';
 import { vitalBreadcrumbs } from '@bodiless/vital-navigation';
 import { vitalImage } from '@bodiless/vital-image';
 import { YouTubeClean, vitalYouTube } from '@bodiless/vital-youtube';
@@ -65,7 +63,7 @@ const Base = asGenericTemplateToken({
       withDesign({
         Image: on(Img)(vitalImage.Hero),
         Video: on(YouTubeClean)(vitalYouTube.Hero),
-        HeroCard: on(CardStatic)(vitalCardStatic.Hero),
+        HeroCard: on(CardStatic)(vitalCardStatic.HeroLeftImageContentCentered),
       }),
     ),
     Content: as(vitalFlowContainer.Default),
@@ -126,26 +124,11 @@ const Generic = asGenericTemplateToken({
   },
 });
 
-const Search = asGenericTemplateToken({
-  ...Base,
-  Meta: {
-    title: 'Search',
-  },
-  Components: {
-    ...Base.Components,
-    Breadcrumb: as(Base.Components.Breadcrumb, addProps({ children: 'Search', })),
-    TopContent: replaceWith(Fragment),
-    Content: on(SearchLayoutClean)(vitalSearchLayout.Default),
-    BottomContent: replaceWith(Fragment),
-  }
-});
-
 interface VitalGenericTemplate extends TokenCollection<GenericTemplateComponents, {}> {
   Base: GenericTemplateToken,
   Generic: GenericTemplateToken,
   ContentListing: GenericTemplateToken,
   WithNoBreadcrumbsOnHomePage: GenericTemplateToken,
-  Search: GenericTemplateToken,
 }
 
 const vitalGenericTemplate: VitalGenericTemplate = {
@@ -153,7 +136,6 @@ const vitalGenericTemplate: VitalGenericTemplate = {
   Generic,
   ContentListing,
   WithNoBreadcrumbsOnHomePage,
-  Search,
 };
 
 export default vitalGenericTemplate;