+++ CHANGELOG.md
@@ -3,6 +3,85 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-rc.33](https://github.com/johnsonandjohnson/bodiless-js/compare/v1.0.0-rc.32...v1.0.0-rc.33) (2023-04-06)
+
+
+### Features
+
+* **components:** 'Editable' passes through html props. [Canary] ([#2020](https://github.com/johnsonandjohnson/bodiless-js/issues/2020)) ([3897084](https://github.com/johnsonandjohnson/bodiless-js/commit/38970844d74ba647a1803f3ef94e4e4dd22239c4))
+
+
+
+
+
+# [1.0.0-rc.32](https://github.com/johnsonandjohnson/bodiless-js/compare/v1.0.0-rc.31...v1.0.0-rc.32) (2023-04-06)
+
+**Note:** Version bump only for package bodiless-js
+
+
+
+
+
+# [1.0.0-rc.31](https://github.com/johnsonandjohnson/bodiless-js/compare/v1.0.0-rc.30...v1.0.0-rc.31) (2023-04-06)
+
+
+### Bug Fixes
+
+* **@bodiless/search, gatsby-theme-bodiless:** Fix/transpile to esm search and gatsby theme bodiless ([#2023](https://github.com/johnsonandjohnson/bodiless-js/issues/2023)) ([ad864af](https://github.com/johnsonandjohnson/bodiless-js/commit/ad864af6f0ef30b8df91528d7e9d4adc6615fbf9))
+
+
+
+
+
+# [1.0.0-rc.30](https://github.com/johnsonandjohnson/bodiless-js/compare/v1.0.0-rc.29...v1.0.0-rc.30) (2023-04-06)
+
+
+### Features
+
+* **gatsby-theme-bodiless, backend, hydration, layouts, tokens, ga4:** Replace crypto-browserify with crypto-js ([#1955](https://github.com/johnsonandjohnson/bodiless-js/issues/1955)) ([bb211c7](https://github.com/johnsonandjohnson/bodiless-js/commit/bb211c74a9e6a1e2e148a5b0d2e5a380250553fe))
+
+
+
+
+
+# [1.0.0-rc.29](https://github.com/johnsonandjohnson/bodiless-js/compare/v1.0.0-rc.28...v1.0.0-rc.29) (2023-03-21)
+
+**Note:** Version bump only for package bodiless-js
+
+
+
+
+
+# [1.0.0-rc.28](https://github.com/johnsonandjohnson/bodiless-js/compare/v1.0.0-rc.27...v1.0.0-rc.28) (2023-03-20)
+
+
+### Bug Fixes
+
+* **styleguide:** accordion & buttons & video ([#2007](https://github.com/johnsonandjohnson/bodiless-js/issues/2007)) ([0e078dc](https://github.com/johnsonandjohnson/bodiless-js/commit/0e078dc605912be5240bca5da7a2b52ba2859cab))
+
+
+
+
+
+# [1.0.0-rc.27](https://github.com/johnsonandjohnson/bodiless-js/compare/v1.0.0-rc.26...v1.0.0-rc.27) (2023-03-16)
+
+
+### Bug Fixes
+
+* **styleguide:** Swap Styleguide for Layout & Typography ([#2003](https://github.com/johnsonandjohnson/bodiless-js/issues/2003)) ([d24e9b1](https://github.com/johnsonandjohnson/bodiless-js/commit/d24e9b1123e36e49d850f755fc5ecccf1f1f92b1))
+
+
+
+
+
+# [1.0.0-rc.26](https://github.com/johnsonandjohnson/bodiless-js/compare/v1.0.0-rc.25...v1.0.0-rc.26) (2023-03-14)
+
+**Note:** Version bump only for package bodiless-js
+
+
+
+
+
 # [1.0.0-rc.25](https://github.com/johnsonandjohnson/bodiless-js/compare/v1.0.0-rc.24...v1.0.0-rc.25) (2023-03-01)