+++ packages/bodiless-backend/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-rc.33](https://github.com/johnsonandjohnson/Bodiless-JS/compare/v1.0.0-rc.32...v1.0.0-rc.33) (2023-04-06)
+
+**Note:** Version bump only for package @bodiless/backend
+
+
+
+
+
+# [1.0.0-rc.32](https://github.com/johnsonandjohnson/Bodiless-JS/compare/v1.0.0-rc.31...v1.0.0-rc.32) (2023-04-06)
+
+**Note:** Version bump only for package @bodiless/backend
+
+
+
+
+
+# [1.0.0-rc.31](https://github.com/johnsonandjohnson/Bodiless-JS/compare/v1.0.0-rc.30...v1.0.0-rc.31) (2023-04-06)
+
+**Note:** Version bump only for package @bodiless/backend
+
+
+
+
+
+# [1.0.0-rc.30](https://github.com/johnsonandjohnson/Bodiless-JS/compare/v1.0.0-rc.29...v1.0.0-rc.30) (2023-04-06)
+
+
+### Features
+
+* **gatsby-theme-bodiless, backend, hydration, layouts, tokens, ga4:** Replace crypto-browserify with crypto-js ([#1955](https://github.com/johnsonandjohnson/Bodiless-JS/issues/1955)) ([bb211c7](https://github.com/johnsonandjohnson/Bodiless-JS/commit/bb211c74a9e6a1e2e148a5b0d2e5a380250553fe))
+
+
+
+
+
+# [1.0.0-rc.29](https://github.com/johnsonandjohnson/Bodiless-JS/compare/v1.0.0-rc.28...v1.0.0-rc.29) (2023-03-21)
+
+**Note:** Version bump only for package @bodiless/backend
+
+
+
+
+
+# [1.0.0-rc.28](https://github.com/johnsonandjohnson/Bodiless-JS/compare/v1.0.0-rc.27...v1.0.0-rc.28) (2023-03-20)
+
+**Note:** Version bump only for package @bodiless/backend
+
+
+
+
+
+# [1.0.0-rc.27](https://github.com/johnsonandjohnson/Bodiless-JS/compare/v1.0.0-rc.26...v1.0.0-rc.27) (2023-03-16)
+
+**Note:** Version bump only for package @bodiless/backend
+
+
+
+
+
+# [1.0.0-rc.26](https://github.com/johnsonandjohnson/Bodiless-JS/compare/v1.0.0-rc.25...v1.0.0-rc.26) (2023-03-14)
+
+**Note:** Version bump only for package @bodiless/backend
+
+
+
+
+
 # [1.0.0-rc.25](https://github.com/johnsonandjohnson/Bodiless-JS/compare/v1.0.0-rc.24...v1.0.0-rc.25) (2023-03-01)
 
 **Note:** Version bump only for package @bodiless/backend